requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyspark>=3.5.0
markdown>=3.5
